DirtySanta fixup v3.1, values freely modifiable via SysFS
When originally implemented, the SysFS interface for the DirtySanta modem
fixup only had the values writeable if they hadn't been specified on the
kernel command-line.  This is actually needlessly paranoid, the interface
is unlikely to be targeted by malicious tools and it is unlikely to fail.
More importantly this is valuable for experimental attempts to load these
values after kernel start.

Change-Id: Ie3d1010ce8c27467323a875c2febd1b277fcc17a